feat: useToggleContents 토글 기능 추가 및 훅 이름 변경
